Moved Subscription logic to Domain/Subscription
