+++ src/Backend/Modules/Mailmotor/Resources/config/forms.yml
@@ -6,13 +6,13 @@ services:
     tags:
         - {name: 'form.type'}
   mailmotor.form.subscription:
-    class: Frontend\Modules\Mailmotor\Form\SubscribeType
+    class: Frontend\Modules\Mailmotor\Domain\Subscription\SubscribeType
     arguments:
       - "@mailmotor.subscriber"
       - "@fork.settings"
     tags:
       - {name: 'form.type'}
   mailmotor.form.unsubscription:
-    class: Frontend\Modules\Mailmotor\Form\UnsubscribeType
+    class: Frontend\Modules\Mailmotor\Domain\Subscription\UnsubscribeType
     tags:
       - {name: 'form.type'}
