+++ src/Frontend/Modules/Mailmotor/Actions/Unsubscribe.php
@@ -12,9 +12,9 @@ namespace Frontend\Modules\Mailmotor\Actions;
 use Frontend\Core\Engine\Base\Block as FrontendBaseBlock;
 use Frontend\Core\Engine\Navigation as FrontendNavigation;
 use Frontend\Core\Language\Locale;
-use Frontend\Modules\Mailmotor\Command\Unsubscription;
-use Frontend\Modules\Mailmotor\Event\NotImplementedUnsubscribedEvent;
-use Frontend\Modules\Mailmotor\Form\UnsubscribeType;
+use Frontend\Modules\Mailmotor\Domain\Subscription\Command\Unsubscription;
+use Frontend\Modules\Mailmotor\Domain\Subscription\Event\NotImplementedUnsubscribedEvent;
+use Frontend\Modules\Mailmotor\Domain\Subscription\UnsubscribeType;
 use MailMotor\Bundle\MailMotorBundle\Exception\NotImplementedException;
 
 /**
