+++ src/Frontend/Modules/Mailmotor/Domain/Subscription/Command/Unsubscription.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Frontend\Modules\Mailmotor\Command;
+namespace Frontend\Modules\Mailmotor\Domain\Subscription\Command;
 
 /*
  * This file is part of the Fork CMS Mailmotor Module from SIESQO.
@@ -11,7 +11,7 @@ namespace Frontend\Modules\Mailmotor\Command;
 
 use Frontend\Core\Language\Locale;
 use Symfony\Component\Validator\Constraints as Assert;
-use Frontend\Modules\Mailmotor\Validator\Constraints as MailingListAssert;
+use Frontend\Modules\Mailmotor\Domain\Subscription\Validator\Constraints as MailingListAssert;
 
 final class Unsubscription
 {
