+++ src/Frontend/Modules/Mailmotor/Domain/Subscription/Event/NotImplementedUnsubscribedEvent.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Frontend\Modules\Mailmotor\Event;
+namespace Frontend\Modules\Mailmotor\Domain\Subscription\Event;
 
-use Frontend\Modules\Mailmotor\Command\Unsubscription;
+use Frontend\Modules\Mailmotor\Domain\Subscription\Command\Unsubscription;
 use Symfony\Component\EventDispatcher\Event;
 
 final class NotImplementedUnsubscribedEvent extends Event
