+++ src/Frontend/Modules/Mailmotor/EventListener/NewNotImplementedMailingListSubscription.php
@@ -11,7 +11,7 @@ namespace Frontend\Modules\Mailmotor\EventListener;
 
 use Common\Language;
 use Common\Mailer\Message;
-use Frontend\Modules\Mailmotor\Event\NotImplementedSubscribedEvent;
+use Frontend\Modules\Mailmotor\Domain\Subscription\Event\NotImplementedSubscribedEvent;
 use Swift_Mailer;
 use Common\ModulesSettings;
