+++ src/Frontend/Modules/Mailmotor/EventListener/NewNotImplementedMailingListUnsubscription.php
@@ -11,7 +11,7 @@ namespace Frontend\Modules\Mailmotor\EventListener;
 
 use Common\Language;
 use Common\Mailer\Message;
-use Frontend\Modules\Mailmotor\Event\NotImplementedUnsubscribedEvent;
+use Frontend\Modules\Mailmotor\Domain\Subscription\Event\NotImplementedUnsubscribedEvent;
 use Swift_Mailer;
 use Common\ModulesSettings;
